+++ src/mass_driver/activity_run.py
@@ -2,6 +2,7 @@
 
 import logging
 import traceback
+from concurrent import futures
 from copy import deepcopy
 from pathlib import Path
 
@@ -51,42 +52,35 @@ def run(
     if migration is not None:
         what_array.append(f"{migration.driver=}")
         patch_results = {}
+
     logger.info(f"Processing {repo_count} with {' and '.join(what_array)}")
-    for repo_index, (repo_id, repo) in enumerate(repos.items(), start=1):
-        repo_logger_name = f"repo.{repo_id}"
-        repo_logger = logging.getLogger(repo_logger_name)
-        try:
-            logger.info(f"[{repo_index:03d}/{repo_count:03d}] Processing {repo_id}...")
-            cloned_repo, repo_gitobj = clone_repo(
-                repo, cache_folder, logger=repo_logger
+
+    futures_map = {}
+    with futures.ThreadPoolExecutor(max_workers=8) as executor:
+        for repo_id, repo in repos.items():
+            repo_logger_name = f"repo.{repo_id}"
+            repo_logger = logging.getLogger(repo_logger_name)
+
+            future_obj = executor.submit(
+                per_repo_process,
+                repo_id,
+                repo,
+                activity,
+                repo_logger,
+                cache_folder,
+                repo_count,
             )
+            futures_map[future_obj] = repo_id
+        # Submitted the jobs: iterate on completion
+        for repo_index, future in enumerate(futures.as_completed(futures_map), start=1):
+            repo_id = futures_map[future]
+            cloned_repo, scan_result, patch_result = future.result()
+            logger.info(f"[{repo_index:04d}/{repo_count:04d}] Processed {repo_id}")
             cloned_repos[repo_id] = cloned_repo
-        except Exception as e:
-            repo_logger.info(f"Error cloning repo '{repo_id}'\nError was: {e}")
-            continue
-        if scan and scanner_results is not None:
-            try:
-                scan_result = scan_repo(scan, cloned_repo)
+            if scanner_results is not None:
                 scanner_results[repo_id] = scan_result
-            except Exception as e:
-                repo_logger.error(f"Error scanning repo '{repo_id}'")
-                repo_logger.error(f"Error was: {e}")
-                # Reaching here should be impossible (catch-all in scan)
-        if migration and patch_results is not None:
-            try:
-                # Ensure no driver persistence between repos
-                migration_copy = deepcopy(migration)
-                result, excep = migrate_repo(
-                    cloned_repo, repo_gitobj, migration_copy, logger=repo_logger
-                )
-                patch_results[repo_id] = result
-            except Exception as e:
-                repo_logger.error(f"Error migrating repo '{repo_id}'")
-                repo_logger.error(f"Error was: {e}")
-                patch_results[repo_id] = PatchResult(
-                    outcome=PatchOutcome.PATCH_ERROR,
-                    details=f"Unhandled exception caught during patching. Error was: {e}",
-                )
+            if patch_results is not None:
+                patch_results[repo_id] = patch_result
     logger.info("Action completed: exiting")
     return ActivityOutcome(
         repos_sourced=repos,
@@ -94,6 +88,40 @@ def run(
         scan_result=scanner_results,
         migration_result=patch_results,
     )
+
+
+def per_repo_process(repo_id, repo, activity, logger, cache_folder, repo_count):
+    """Process a single repo, in-thread"""
+    try:
+        logger.info(f"Processing {repo_id}...")
+        cloned_repo, repo_gitobj = clone_repo(repo, cache_folder, logger=logger)
+    except Exception as e:
+        logger.info(f"Error cloning repo '{repo_id}'\nError was: {e}")
+        raise e  # FIXME: Use custom exeption for capturing error here
+    scan_result: ScanResult | None = None
+    if activity.scan is not None:
+        try:
+            scan_result = scan_repo(activity.scan, cloned_repo)
+        except Exception as e:
+            logger.error(f"Error scanning repo '{repo_id}'")
+            logger.error(f"Error was: {e}")
+            # Reaching here should be impossible (catch-all in scan)
+    patch_result: PatchResult | None = None
+    if activity.migration:
+        try:
+            # Ensure no driver persistence between repos
+            migration_copy = deepcopy(activity.migration)
+            patch_result, excep = migrate_repo(
+                cloned_repo, repo_gitobj, migration_copy, logger=logger
+            )
+        except Exception as e:
+            logger.error(f"Error migrating repo '{repo_id}'")
+            logger.error(f"Error was: {e}")
+            patch_result = PatchResult(
+                outcome=PatchOutcome.PATCH_ERROR,
+                details=f"Unhandled exception caught during patching. Error was: {e}",
+            )  # FIXME: Catch custom-exception into the PatchResult object
+    return (cloned_repo, scan_result, patch_result)
 
 
 def clone_repo(
@@ -111,7 +139,7 @@ def clone_repo(
     return cloned_repo, repo_gitobj
 
 
-# TODO: Avoid passing out the exception, catch the trace in details kw (see scanner_run)
+# FIXME: Throw custom exception on failure
 def migrate_repo(
     cloned_repo: ClonedRepo,
     repo_gitobj: GitRepo,